+++ spec/project/project_helper_spec.rb
@@ -54,9 +54,15 @@ module ProjectSpecs
     #-------------------------------------------------------------------------#
 
     describe "System Frameworks" do
+
+      before do
+        @target = Xcodeproj::Project::ProjectHelper.new_target(@project, :static_library, 'Pods', :ios, '6.0', @project.products_group)
+        Xcodeproj::XcodebuildHelper.any_instance.stubs(:last_ios_sdk).returns(Xcodeproj::Constants::LAST_KNOWN_IOS_SDK)
+      end
+
       it "adds a file reference for a system framework, to the Frameworks group" do
-        target = stub(:sdk => 'iphoneos5.0')
-        file = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'QuartzCore', target)
+        @target.stubs(:sdk).returns('iphoneos5.0')
+        file = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'QuartzCore', @target)
         file.parent.should == @project['Frameworks']
         file.name.should == 'QuartzCore.framework'
         file.path.should.match %r|Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS5.0.sdk/System/Library/Frameworks/QuartzCore.framework|
@@ -64,20 +70,24 @@ module ProjectSpecs
       end
 
       it "links system frameworks to the last known SDK if needed" do
-        target = stub(:sdk => 'iphoneos')
-        file = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'QuartzCore', target)
+        @target.stubs(:sdk).returns('iphoneos')
+        file = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'QuartzCore', @target)
         file.path.should.match %r|Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS.*.sdk/System/Library/Frameworks/QuartzCore.framework|
         file.source_tree.should == 'DEVELOPER_DIR'
       end
 
       it "does not add a system framework if it already exists in the project" do
-        target = stub(:sdk => 'iphoneos6.0')
-        file_1 = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'Foundation', target)
+        file_1 = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'Foundation', @target)
         file_1.name.should == 'Foundation.framework'
         before_size = @project.frameworks_group.files.size
-        file_2 = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'Foundation', target)
+        file_2 = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'Foundation', @target)
         file_2.should == file_1
         @project.frameworks_group.files.size.should == before_size
+      end
+
+      it "adds the framework to the framework build phase of the target" do
+        ref = Xcodeproj::Project::ProjectHelper.add_system_framework(@project, 'QuartzCore', @target)
+        @target.frameworks_build_phase.files_references.should.include(ref)
       end
     end